+++ packages/fcl/src/serialize/index.js
@@ -4,9 +4,11 @@ import {config, createSignableVoucher} from "@onflow/sdk"
 
 export const serialize = async (args = [], opts = {}) => {
   // prettier-ignore
-  const resolveFunction = await config()
-    .get("sdk.resolve", opts.resolve || defaultResolve(opts))
-
+  const resolveFunction = await config().get(
+    "sdk.resolve",
+    opts.resolve || defaultResolve
+  )
+  
   if (Array.isArray(args)) args = await pipe(interaction(), args)
 
   return JSON.stringify(